Extract scale modification to standalone example
