+++ appyx-components/spotlight/common/src/commonMain/kotlin/com/bumble/appyx/components/spotlight/ui/sliderscale/MutableUiState.kt
@@ -1,12 +1,10 @@
-package com.bumble.appyx.components.spotlight.ui.sliderrotation
+package com.bumble.appyx.components.spotlight.ui.sliderscale
 
 import androidx.compose.animation.core.SpringSpec
 import androidx.compose.animation.core.spring
 import androidx.compose.ui.Modifier
 import com.bumble.appyx.interactions.core.ui.context.UiContext
-import com.bumble.appyx.interactions.core.ui.property.impl.Alpha
 import com.bumble.appyx.interactions.core.ui.property.impl.Position
-import com.bumble.appyx.interactions.core.ui.property.impl.RotationY
 import com.bumble.appyx.interactions.core.ui.property.impl.Scale
 import com.bumble.appyx.interactions.core.ui.state.BaseMutableUiState
 import kotlinx.coroutines.CoroutineScope
@@ -18,24 +16,18 @@ class MutableUiState(
     uiContext: UiContext,
     val position: Position,
     val scale: Scale,
-    val rotationY: RotationY,
-    val alpha: Alpha,
 ) : BaseMutableUiState<MutableUiState, TargetUiState>(
     uiContext = uiContext,
-    motionProperties = listOf(scale, alpha, position)
+    motionProperties = listOf(scale, position)
 ) {
     override val modifier: Modifier
         get() = Modifier.then(position.modifier)
             .then(scale.modifier)
-            .then(rotationY.modifier)
-            .then(alpha.modifier)
 
     override suspend fun snapTo(scope: CoroutineScope, target: TargetUiState) {
         scope.launch {
             position.snapTo(target.position.value)
             scale.snapTo(target.scale.value)
-            rotationY.snapTo(target.rotationY.value)
-            alpha.snapTo(target.alpha.value)
         }
     }
 
@@ -55,14 +47,6 @@ class MutableUiState(
                         target.scale.value,
                         spring(springSpec.dampingRatio, springSpec.stiffness)
                     )
-                    rotationY.animateTo(
-                        target.rotationY.value,
-                        spring(springSpec.dampingRatio, springSpec.stiffness)
-                    )
-                    alpha.animateTo(
-                        target.alpha.value,
-                        spring(springSpec.dampingRatio, springSpec.stiffness)
-                    )
                 }
             ).awaitAll()
         }
@@ -72,8 +56,6 @@ class MutableUiState(
         scope.launch {
             position.lerpTo(start.position, end.position, fraction)
             scale.lerpTo(start.scale, end.scale, fraction)
-            rotationY.lerpTo(start.rotationY, end.rotationY, fraction)
-            alpha.lerpTo(start.alpha, end.alpha, fraction)
         }
     }
 }
